+++ docs/architecture/c-sharp-client-library.md
