+++ samples/azure-function/GetForecast.cs
@@ -1,35 +1,36 @@
-using System;
-using System.IO;
-using System.Threading.Tasks;
+using GSF.CarbonAware.Handlers;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
-using GSF.CarbonAware.Handlers;
+using System;
+using System.IO;
+using System.Threading.Tasks;
+using System.Text.Json;
 
-namespace function
+
+namespace CarbonAwareFunctions
 {
-    public class GetCarbonIntensity
+    public class GetForecast
     {
-        private readonly IEmissionsHandler _handler;
+        private readonly IForecastHandler _handler;
 
-        public GetCarbonIntensity(IEmissionsHandler handler)
+        public GetForecast(IForecastHandler handler)
         {
             this._handler = handler;
         }
 
-
-        [FunctionName("GetAverageCarbonIntensity")]
+        [FunctionName("GetCurrentForecast")]
         public async Task<IActionResult> Run(
             [HttpTrigger(AuthorizationLevel.Function, "get", "post", Route = null)] HttpRequest req,
             ILogger log)
         {
-            //Get the startDate, endDate, and location from the request query if the values are present in the query
+            //Get the startDate, endDate, location, and duration from the request query if the values are present in the query
             string startDate = req.Query["startdate"];
             string endDate = req.Query["enddate"];
             string location = req.Query["location"];
+            string duration = req.Query["duration"];
 
             //If the parameters are includes in the body of the request, read the values from the request body
             string requestBody = String.Empty;
@@ -37,17 +38,16 @@ namespace function
             {
                 requestBody = await streamReader.ReadToEndAsync();
             }
-            dynamic data = JsonConvert.DeserializeObject(requestBody);
+            dynamic data = JsonSerializer.Deserialize<object>(requestBody);
 
             startDate ??= data?.startDate;
             endDate ??= data?.endDate;
             location ??= data?.location;
-
+            duration ??= data?.duration;
 
             try
             {
-                var result = await _handler.GetAverageCarbonIntensityAsync(location, DateTimeOffset.Parse(startDate), DateTimeOffset.Parse(endDate));
-                log.LogInformation($"For location {location} Starting at: {startDate} Ending at: {endDate} the Average Emissions Rating is: {result}.");
+                var result = await _handler.GetCurrentForecastAsync(new string[] { location }, DateTimeOffset.Parse(startDate), DateTimeOffset.Parse(endDate), int.Parse(duration));
 
                 return new OkObjectResult(result);
             }
@@ -55,9 +55,9 @@ namespace function
             {
                 //Messages related to incorrect parameter values (ie dates outside of range) are returned in the data section
                 //Otherwise send the returned error messages
-                if (e.Data.Count>0)
+                if (e.Data.Count > 0)
                     return new BadRequestObjectResult(e.Data);
-                else 
+                else
                     return new BadRequestObjectResult(e.Message);
             }
