+++ samples/azure-function/README.md
@@ -1,12 +1,12 @@
 # Use Carbon Aware SDK with an Azure Function
 
 ## Overview
-The two samples included showcase the Azure Functions tooling for the Carbon Aware SDK [C# Class Library](../../docs/architecture/c%23-client-library.md). The emissions function app implements GetAverageCarbonIntensity and the forecast function app implements GetCurrentForecast. GetAverageCarbonIntensity uses the EmissionsHandler to return the Carbon Intensity rate of a location for a specific timespan. GetCurrentForecast uses the ForecastHandler to yield the optimal time of a specified location and duration. The functions can run locally for debugging or be deployed to Azure.
+
+The sample included showcase the Azure Functions tooling for the Carbon Aware SDK [C# Class Library](../../docs/architecture/c-sharp-client-library.md). It includes an implementation for `GetAverageCarbonIntensity` and for `GetCurrentForecast`. `GetAverageCarbonIntensity` uses the `EmissionsHandler` to return the carbon intensity rate of a location for a specific timespan. `GetCurrentForecast` uses the `ForecastHandler` to yield the optimal time of a specified location and duration. The functions can run locally for debugging or be deployed to Azure.
 
 ## Azure Function Dependency Injection
-The Carbon Aware SDK is included in the function .csproj file by [creating and adding the SDK as a package](../../docs/packaging.md#included-scripts).  The [Startup.cs](Startup.cs) file uses dependency injection to access the handlers in the library. The following code initializes the C# Library:
 
-_For the emissions function app:_
+The Carbon Aware SDK is included in the function .csproj file by [creating and adding the SDK as a package](../../docs/packaging.md#included-scripts).  The [Startup.cs](./Startup.cs) file uses dependency injection to access the handlers in the library. The following code initializes the C# Library:
 
 ```C#
  public override void Configure(IFunctionsHostBuilder builder)
@@ -14,48 +14,41 @@ _For the emissions function app:_
             var configuration = builder.GetContext().Configuration;
             builder.Services
             .AddEmissionsServices(configuration)
-        }
-```
-
-_For the forecast function app:_
-```C#
- public override void Configure(IFunctionsHostBuilder builder)
-        {
-            var configuration = builder.GetContext().Configuration;
-            builder.Services
             .AddForecastServices(configuration);
         }
 ```
 
-
 ## Run Function Locally
-Both azure function apps can be run locally without needing an azure subscription. The process for running both is the same, as they use the same configuration, just call different paths within the SDK.
 
-### Prerequisites
+Both Azure Function apps can be run locally without needing an azure subscription. The process for running both is the same, as they use the same configuration, just call different paths within the SDK.
+
+### Prerequisites to Run
+
 [.NET Core SDK](https://dotnet.microsoft.com/download)
 
 [Azure Functions Core Tools](https://learn.microsoft.com/en-us/azure/azure-functions/functions-run-local)
 
 ### Start Function
+
 To run and debug locally, update the [appsettings.json](appsettings.json) file to include the desired [configuration](../../docs/configuration.md).
 
-In the app folder (`samples/azure-function/emissions-azure-function` or `samples/azure-function/forecast-azure-function`), run the command:  ```func start```
+In the app folder (`samples/azure-function`), run the command:  ```func start```
 
 After the function has compiled and is running, the URLs to the functions will be presented.  
 
-_Example call for Emissions azure function_
+#### _Example call for Get Average Carbon Intensity function_
 
 The following example will retrieve the Average Carbon Intensity.  For this example, query parameters were used, but the values could also be sent in the body of the request.
 
-```
+```bash
 curl --location --request GET 'http://localhost:7071/api/GetAverageCarbonIntensity?startDate=2022-03-01T15:30:00Z&endDate=2022-03-01T18:30:00Z&location=eastus'
 ```
 
-_Example call for Forecast azure function_
+#### _Example call for Get Current Forecast function_
 
 The following example will call the Current Forecast route.  If an error is returned, update the start and end dates.  The request can use either the request body or query parameters.
 
-```
+```bash
 curl --location --request GET 'http://localhost:7071/api/GetCurrentForecast' \
 --header 'Content-Type: application/json' \
 --data-raw '{
@@ -67,20 +60,22 @@ curl --location --request GET 'http://localhost:7071/api/GetCurrentForecast' \
 ```
 
 ## Deploy to Azure
+
 If you have an azure subscription, you can also deploy these functions to Azure.
 
-### Prerequisites
+### Prerequisites to Deploy
 
 You must have the [Azure CLI](https://learn.microsoft.com/en-us/cli/azure/install-azure-cli) or [Azure PowerShell](https://learn.microsoft.com/en-us/powershell/azure/install-az-ps) installed locally to be able to publish to Azure.
 
 [Azure Functions Core Tools](https://learn.microsoft.com/en-us/azure/azure-functions/functions-run-local)
 
 ### Create Function App
+
 Log in to Azure:  ```az login```
 
 Once the correct subscription is selected run the following script to create a new function app:
 
-```
+```bash
 # Function app and storage account names must be unique.
 
 # Variable block
@@ -105,13 +100,14 @@ echo "Creating $functionApp"
 az functionapp create --name $functionApp --storage-account $storage --consumption-plan-location "$location" --resource-group $resourceGroup --functions-version $functionsVersion
 ```
 
-### Publish Functions 
-Update the [appsettings.json](appsettings.json) file to include the desired [configuration](../../docs/configuration.md).
+### Publish Functions
+
+Update the [appsettings.json](./appsettings.json) file to include the desired [configuration](../../docs/configuration.md).
 
 To publish the function code to a function app in Azure, use the publish command in the samples/azure-function folder:
 
-```
-func azure functionapp publish $functionApp
+```bash
+func Azure Functionapp publish $functionApp
 ```
 
 ## References
