--- samples/azure-function/emissions-azure-function/.devcontainer/devcontainer.json
@@ -1,5 +1,0 @@
-{
-	"name": "Azure Functions & C# - .NET 6 (In-Process)",
-	"dockerFile": "../Dockerfile",
-	"forwardPorts": [ 7071 ],
-}
