--- samples/azure-function/emissions-azure-function/Startup.cs
@@ -1,30 +1,0 @@
-using System.IO;
-using GSF.CarbonAware.Configuration;
-using Microsoft.Azure.Functions.Extensions.DependencyInjection;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-[assembly: FunctionsStartup(typeof(CarbonAwareFunctions.Startup))]
-
-namespace CarbonAwareFunctions
-{
-    public class Startup : FunctionsStartup
-    {
-        public override void ConfigureAppConfiguration(IFunctionsConfigurationBuilder builder)
-        {
-            FunctionsHostBuilderContext context = builder.GetContext();
-
-            builder.ConfigurationBuilder
-                .AddJsonFile(Path.Combine(context.ApplicationRootPath, "appsettings.json"), optional: true, reloadOnChange: false)
-                .AddJsonFile(Path.Combine(context.ApplicationRootPath, $"appsettings.{context.EnvironmentName}.json"), optional: true, reloadOnChange: false)
-                .AddEnvironmentVariables();
-        }
-
-        public override void Configure(IFunctionsHostBuilder builder)
-        {
-            var configuration = builder.GetContext().Configuration;
-            builder.Services
-            .AddEmissionsServices(configuration);
-        }
-    }
-}
