--- samples/azure-function/forecast-azure-function/appsettings.json
@@ -1,15 +1,0 @@
-{
-  "DataSources": {
-    "EmissionsDataSource": "",
-    "ForecastDataSource": "WattTime",
-    "Configurations": {
-      "WattTime": {
-        "Type": "WattTime",
-        "Username": "username",
-        "Password": "password",
-        "BaseURL": "https://api2.watttime.org/v2/"
-      }
-    }
-  }
-}
-
