--- samples/azure-function/forecast-azure-function/host.json
@@ -1,11 +1,0 @@
-{
-    "version": "2.0",
-    "logging": {
-        "applicationInsights": {
-            "samplingSettings": {
-                "isEnabled": true,
-                "excludedTypes": "Request"
-            }
-        }
-    }
-}
